Change app global state to its own context folder
