+++ src/components/EmojiGuide/EmojiGuide.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Link, Redirect } from 'react-router-dom'
-import {useAuth} from '../../contexts/AuthContext'
+import { useAuth } from '../../contexts/AuthContext'
 import emojiData from './emojiData'
 
 import {
@@ -17,7 +17,7 @@ import {
 } from 'react95'
 
 const EmojiGuide = () => {
-  const {currentUser} = useAuth()
+  const { currentUser } = useAuth()
   const data = emojiData.map(emoji => {
     return (
       <TableRow>
