+++ src/components/ResultsDisplay/ResultsDisplay.js
@@ -1,6 +1,7 @@
 import React, { useState, useEffect } from 'react' 
 import { Link, Redirect } from 'react-router-dom'
 import { useAuth } from '../../contexts/AuthContext'
+import { useApp } from '../../contexts/AppContext'
 import TabDisplay from '../TabDisplay/TabDisplay'
 
 import './ResultsDisplay.css'
@@ -17,7 +18,9 @@ import {
 
 const ResultsDisplay = () => {
   const [activeTab, setActiveTab] = useState(0)
-  const { termData, termName, resetSearchData, currentUser } = useAuth()
+  const { currentUser } = useAuth()
+  const { termData, termName, resetSearchData } = useApp()
+
   const displayTermName = termName.replace(/%20/g, " ")
 
   const definitionTabs = termData.map((term, i) => <Tab value={i} key={i} id={i}>{`Definition ${i+1}`}</Tab>)
