+++ src/contexts/AuthContext.js
@@ -1,5 +1,4 @@
 import React, { useContext, useState, useEffect } from 'react'
-import apiCalls from './apiCalls'
 import { auth, googleProvider, twitterProvider, githubProvider } from './firebase'
 
 const Context = React.createContext()
@@ -11,10 +10,6 @@ export function useAuth() {
 export default function AuthProvider({ children }) {
   const [currentUser, setCurrentUser] = useState()
   const [displayName, setDisplayName] = useState('User')
-  const [termName, setTermName] = useState('')
-  const [termData, setTermData] = useState([])
-  const [userFavorites, setUserFavorites ] = useState([])
-  const [displayTheme, setDisplayTheme] = useState('original')
 
   function signUp(email, password) {
     return auth.createUserWithEmailAndPassword(email, password)
@@ -57,48 +52,6 @@ export default function AuthProvider({ children }) {
     setDisplayName(name)
   }
 
-  function sortIncomingData(data) {
-    const results = data.sort((a,b) => b.thumbs_up - a.thumbs_up)
-    return [results[0], results[1], results[2]]
-  }
-
-  async function querySearchTerms(terms) {
-    setTermName(terms)
-    
-    try {
-     const result = await apiCalls.requestTermsInfo(terms)
-     const data = await result.list
-     const sortedData = await sortIncomingData(data)
-     await setTermData(sortedData)
-    } catch (error) {
-      console.log(error)
-    }
-  }
-
-  function resetSearchData(){
-    setTermData([])
-    setTermName('')
-  }
-
-  function storeUserFavorites(term){
-    const { definition, word } = term
-    const favorite = {
-      id: userFavorites.length + 1,
-      word,
-      definition,
-    }
-    setUserFavorites([...userFavorites, favorite])
-  }
-
-  function removeFavorite(id){
-    const newFavorites = userFavorites.filter(term => term.id !== id)
-    setUserFavorites(newFavorites)
-  }
-
-  function updateTheme(theme){
-    setDisplayTheme(theme)
-  }
-
   useEffect(() => {
     const changeUserState = auth.onAuthStateChanged(user => {
       if (user) {
@@ -112,20 +65,11 @@ export default function AuthProvider({ children }) {
   const value = {
     currentUser,
     displayName,
-    userFavorites,
     signOut,
     signUp,
     login,
     signInWithPopup,
     updateName,
-    querySearchTerms,
-    storeUserFavorites,
-    removeFavorite,
-    resetSearchData,
-    updateTheme,
-    displayTheme,
-    termData,
-    termName
   }
 
   return (
